requirements.txt
fastapi
uvicorn
langchain-core
langchain-google-genai
python-dotenv
requests
pyttsx3